+++ src/rules/value-keyword-case/index.js
@@ -20,6 +20,13 @@ const ignoredCharacters = new Set([
   "+", "-", "/", "*", "%",
 ])
 
+const ignoredCamelCaseKeywords = {
+  "optimizespeed": "optimizeSpeed",
+  "optimizelegibility": "optimizeLegibility",
+  "geometricprecision": "geometricPrecision",
+  "currentcolor": "currentColor",
+}
+
 export default function (expectation, options) {
   return (root, result) => {
     const validOptions = validateOptions(result, ruleName, {
@@ -51,7 +58,6 @@ export default function (expectation, options) {
           || cssWordIsVariable(node.value)
           || value.indexOf("#") !== -1
           || ignoredCharacters.has(keyword)
-          || keyword === "currentColor"
         ) { return }
 
         const parsedUnit = valueParser.unit(keyword)
@@ -62,7 +68,18 @@ export default function (expectation, options) {
 
         if (ignoreKeywords.length > 0 && matchesStringOrRegExp(keyword, ignoreKeywords)) { return }
 
-        const expectedKeyword = expectation === "lower" ? keyword.toLowerCase() : keyword.toUpperCase()
+        const keywordLowerCase = keyword.toLocaleLowerCase()
+        let expectedKeyword = null
+
+        if (expectation === "lower"
+          && ignoredCamelCaseKeywords.hasOwnProperty(keywordLowerCase)
+        ) {
+          expectedKeyword = ignoredCamelCaseKeywords[keywordLowerCase]
+        } else if (expectation === "lower") {
+          expectedKeyword = keyword.toLowerCase()
+        } else {
+          expectedKeyword = keyword.toUpperCase()
+        }
 
         if (keyword === expectedKeyword) { return }